requirements.txt
kafka-python
faker
pytz
psycopg2
python-dotenv
SQLAlchemy
streamlit
plotly